+++ Apiraiser/Services/UserService.cs
@@ -232,52 +232,48 @@ namespace Apiraiser.Services
 
                 int Id = Int32.Parse(result["Id"]?.ToString() ?? null);
 
-                APIResult role = await ServiceManager.Instance.GetService<UsersService>().GetUserRole(Id);
-                if (role.Success == true && role.Data is not null)
+                APIResult roles = await ServiceManager.Instance.GetService<UsersService>().GetUserRoles(Id);
+                List<int> roleIds = new List<int>();
+                if (roles.Success == true && roles.Data is not null)
                 {
-                    result.Add("Role", (Dictionary<string, object>)(role.Data));
-                    jwtTokenData.Add("Roles", (Dictionary<string, object>)(role.Data));
+                    roleIds = ((List<Dictionary<string, object>>)(roles.Data)).Select(x => (int)x["Id"]).ToList();
+                    result.Add("Roles", roles.Data);
+                    jwtTokenData.Add("Roles", ((List<Dictionary<string, object>>)(roles.Data)));
                 }
 
-                int roleId = Int32.Parse(((Dictionary<string, object>)(role.Data))["Id"].ToString());
+                // APIResult rolePermissionGroupMappings = await ServiceManager.Instance.GetService<UsersService>().GetRolePermissionGroupMappings(roleId);
+                // if (rolePermissionGroupMappings.Success == true && rolePermissionGroupMappings.Data is not null)
+                // {
+                //     result.Add("RolePermissionGroupMappings", (List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data));
+                //     jwtTokenData.Add("RolePermissionGroupMappings", (List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data));
 
+                //     List<int> permissionGroupIds = ((List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data)).Select(x => Int32.Parse(x["PermissionGroup"].ToString())).Distinct().ToList();
 
+                //     APIResult permissionGroupMappings = await ServiceManager.Instance.GetService<UsersService>().GetPermissionGroupMappings(permissionGroupIds);
 
-                APIResult rolePermissionGroupMappings = await ServiceManager.Instance.GetService<UsersService>().GetRolePermissionGroupMappings(roleId);
-                if (rolePermissionGroupMappings.Success == true && rolePermissionGroupMappings.Data is not null)
-                {
-                    result.Add("RolePermissionGroupMappings", (List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data));
-                    jwtTokenData.Add("RolePermissionGroupMappings", (List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data));
+                //     if (permissionGroupMappings.Success == true && permissionGroupMappings.Data is not null)
+                //     {
+                //         result.Add("PermissionGroupMappings", (List<Dictionary<string, object>>)(permissionGroupMappings.Data));
+                //     }
 
-                    List<int> permissionGroupIds = ((List<Dictionary<string, object>>)(rolePermissionGroupMappings.Data)).Select(x => Int32.Parse(x["PermissionGroup"].ToString())).Distinct().ToList();
+                //     APIResult permissionGroups = await ServiceManager.Instance.GetService<PermissionGroupsService>().GetPermissionGroups();
 
-                    APIResult permissionGroupMappings = await ServiceManager.Instance.GetService<UsersService>().GetPermissionGroupMappings(permissionGroupIds);
+                //     if (permissionGroups.Success == true && permissionGroups.Data is not null)
+                //     {
+                //         result.Add("PermissionGroups", ((List<Dictionary<string, object>>)(permissionGroups.Data)).Where(x => permissionGroupIds.Contains((int)x["Id"])).ToList());
+                //     }
 
-                    if (permissionGroupMappings.Success == true && permissionGroupMappings.Data is not null)
-                    {
-                        result.Add("PermissionGroupMappings", (List<Dictionary<string, object>>)(permissionGroupMappings.Data));
-                    }
+                //     List<int> permissionIds = ((List<Dictionary<string, object>>)(permissionGroupMappings.Data))
+                //         .Select(x => Int32.Parse(x["Permission"].ToString()))
+                //         .ToList();
 
-                    APIResult permissionGroups = await ServiceManager.Instance.GetService<PermissionGroupsService>().GetPermissionGroups();
-
-                    if (permissionGroups.Success == true && permissionGroups.Data is not null)
-                    {
-                        result.Add("PermissionGroups", ((List<Dictionary<string, object>>)(permissionGroups.Data)).Where(x => permissionGroupIds.Contains((int)x["Id"])).ToList());
-                    }
-
-                    List<int> permissionIds = ((List<Dictionary<string, object>>)(permissionGroupMappings.Data))
-                        .Select(x => Int32.Parse(x["Permission"].ToString()))
-                        .ToList();
-
-                    APIResult permissions = await ServiceManager.Instance.GetService<UsersService>().GetPermissionsByIds(permissionIds);
-                    if (permissions.Success == true && permissions.Data is not null)
-                    {
-                        result.Add("Permissions", (List<Dictionary<string, object>>)(permissions.Data));
-                        jwtTokenData.Add("Permissions", (List<Dictionary<string, object>>)(permissions.Data));
-                    }
-                }
-
-
+                //     APIResult permissions = await ServiceManager.Instance.GetService<UsersService>().GetPermissionsByIds(permissionIds);
+                //     if (permissions.Success == true && permissions.Data is not null)
+                //     {
+                //         result.Add("Permissions", (List<Dictionary<string, object>>)(permissions.Data));
+                //         jwtTokenData.Add("Permissions", (List<Dictionary<string, object>>)(permissions.Data));
+                //     }
+                // }
 
                 result.Add("Token", JWT.GenerateJSONWebToken(_configuration, jwtTokenData));